python/07-ux-demos/human-in-the-loop-approval-agent/src/pyproject.toml
[project]
name = "agent-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=1.4.7",
    "python-dotenv>=1.2.2",
    "strands-agents>=1.34.1",
]
